+++ Changelog.md
@@ -4,6 +4,7 @@
 
 * Added category field to accounting, uses data from project field
 * Added db:backup console command
+* Show wallet amount in dashboard
 
 ## 2.5.1